requirements.txt
PySide6>=6.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pycryptodome>=3.15.0
pyserial>=3.5
